# docs/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: POC Chatbot Kendra
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local Developnent
  - url: https://gveecfcbg4.execute-api.us-east-1.amazonaws.com/Prod 
    description: AWS Developnent
security:
  - BearerAuth: [ ]
paths:
...
  /auth/register:
    post:
      summary: Registro de usuario
      tags:
        - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Refresh exitoso
...
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
...
  securitySchemes:
    BearerAuth: # Nombre que elijas para el esquema
      type: http
      scheme: bearer
      bearerFormat: JWT